+++ src/main/java/domain/ladderGame/LadderGame.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladderGame;
 
 import domain.ladder.DirectionalPoint;
 import domain.ladder.Ladder;
