+++ src/test/java/domain/ladderGame/LadderGameTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladderGame;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
